Core/World: Change worldstate storage to uint32 to fix compile warnings (its also saved/loaded as uint32 from database)

+++ src/server/game/Battlefield/Zones/BattlefieldTB.cpp
@@ -79,14 +79,14 @@ bool BattlefieldTB::SetupBattlefield()
 
     // Set timer
     m_Timer = sWorld->getWorldState(TB_WS_TIME_NEXT_BATTLE);
-    
+
     // Defending team isn't set yet? Choose randomly.
     if (sWorld->getWorldState(TB_WS_FACTION_CONTROLLING) == 0)
-        sWorld->setWorldState(TB_WS_FACTION_CONTROLLING, uint64(urand(1, 2)));
+        sWorld->setWorldState(TB_WS_FACTION_CONTROLLING, uint32(urand(1, 2)));
 
     // Set defender team
     SetDefenderTeam(TeamId(sWorld->getWorldState(TB_WS_FACTION_CONTROLLING) - 1));
-    
+
     // Just to save world states
     SendInitWorldStatesToAll();
 
@@ -103,7 +103,7 @@ bool BattlefieldTB::SetupBattlefield()
         }
         AddCapturePoint(capturePoint);
     }
-    
+
     // Spawn towers
     for (uint8 i = 0; i < TB_TOWERS_COUNT; i++)
         if (GameObject* go = SpawnGameObject(TBTowers[i].entry, TBTowers[i].x, TBTowers[i].y, TBTowers[i].z, TBTowers[i].o))
@@ -132,7 +132,7 @@ bool BattlefieldTB::SetupBattlefield()
     warnedFiveMinutes = false;
     warnedTwoMinutes = false;
     warnedOneMinute = false;
-    
+
     UpdateNPCsAndGameObjects();
 
     return true;
@@ -163,7 +163,7 @@ bool BattlefieldTB::Update(uint32 diff)
             SendWarning(TB_TEXT_PREPARATIONS_IN_1_MIN);
         }
     }
-    
+
     if (!updatedNPCAndObjects)
     {
         if (m_updateObjectsTimer <= diff)
@@ -299,7 +299,7 @@ void BattlefieldTB::FillInitialWorldStates(WorldPackets::WorldState::InitWorldSt
     packet.Worldstates.emplace_back(uint32(TB_WS_BUILDINGS_CAPTURED_SHOW), int32(IsWarTime() ? 1 : 0));
     packet.Worldstates.emplace_back(uint32(TB_WS_BUILDINGS_CAPTURED), int32(GetData(BATTLEFIELD_TB_DATA_BUILDINGS_CAPTURED)));
     packet.Worldstates.emplace_back(uint32(TB_WS_TOWERS_DESTROYED), int32(0));
-    
+
     packet.Worldstates.emplace_back(uint32(TB_WS_TIME_BATTLE_END_SHOW), int32(IsWarTime() ? 1 : 0));
 
     packet.Worldstates.emplace_back(uint32(TB_WS_STATE_BATTLE), int32(IsWarTime() ? 1 : 0));
@@ -340,7 +340,7 @@ void BattlefieldTB::FillInitialWorldStates(WorldPackets::WorldState::InitWorldSt
     }
 
     packet.Worldstates.emplace_back(uint32(TB_WS_TIME_NEXT_BATTLE_SHOW), int32(!IsWarTime() ? 1 : 0));
-    
+
     packet.Worldstates.emplace_back(uint32(TB_WS_ALLIANCE_CONTROLS_SHOW), int32(!IsWarTime() && GetDefenderTeam() == TEAM_ALLIANCE ? 1 : 0));
     packet.Worldstates.emplace_back(uint32(TB_WS_HORDE_CONTROLS_SHOW), int32(!IsWarTime() && GetDefenderTeam() == TEAM_HORDE ? 1 : 0));
 
@@ -352,7 +352,7 @@ void BattlefieldTB::FillInitialWorldStates(WorldPackets::WorldState::InitWorldSt
     //packet.Worldstates.emplace_back(uint32(TB_WS_66_UNKNOWN), int32(0));
 
     packet.Worldstates.emplace_back(uint32(TB_WS_KEEP_ALLIANCE), int32(GetDefenderTeam() == TEAM_ALLIANCE ? 1 : 0));
-    packet.Worldstates.emplace_back(uint32(TB_WS_KEEP_HORDE), int32(GetDefenderTeam() == TEAM_HORDE ? 1 : 0));    
+    packet.Worldstates.emplace_back(uint32(TB_WS_KEEP_HORDE), int32(GetDefenderTeam() == TEAM_HORDE ? 1 : 0));
 }
 
 void BattlefieldTB::SendInitWorldStatesTo(Player* player)
@@ -370,13 +370,13 @@ void BattlefieldTB::SendInitWorldStatesTo(Player* player)
 void BattlefieldTB::SendInitWorldStatesToAll()
 {
     // Save
-    sWorld->setWorldState(TB_WS_STATE_BATTLE, uint64(IsWarTime() ? 1 : 0));
-    sWorld->setWorldState(TB_WS_ALLIANCE_CONTROLS_SHOW, uint64(!IsWarTime() && GetDefenderTeam() == TEAM_ALLIANCE ? 1 : 0));
-    sWorld->setWorldState(TB_WS_HORDE_CONTROLS_SHOW, uint64(!IsWarTime() && GetDefenderTeam() == TEAM_HORDE ? 1 : 0));
-    sWorld->setWorldState(TB_WS_ALLIANCE_ATTACKING_SHOW, uint64(IsWarTime() && GetAttackerTeam() == TEAM_ALLIANCE ? 1 : 0));
-    sWorld->setWorldState(TB_WS_HORDE_ATTACKING_SHOW, uint64(IsWarTime() && GetAttackerTeam() == TEAM_HORDE ? 1 : 0));
-    sWorld->setWorldState(TB_WS_TIME_NEXT_BATTLE, uint64(!IsWarTime() ? m_Timer : 0));
-    sWorld->setWorldState(TB_WS_TIME_NEXT_BATTLE_SHOW, uint64(!IsWarTime() ? 1 : 0));
+    sWorld->setWorldState(TB_WS_STATE_BATTLE, uint32(IsWarTime() ? 1 : 0));
+    sWorld->setWorldState(TB_WS_ALLIANCE_CONTROLS_SHOW, uint32(!IsWarTime() && GetDefenderTeam() == TEAM_ALLIANCE ? 1 : 0));
+    sWorld->setWorldState(TB_WS_HORDE_CONTROLS_SHOW, uint32(!IsWarTime() && GetDefenderTeam() == TEAM_HORDE ? 1 : 0));
+    sWorld->setWorldState(TB_WS_ALLIANCE_ATTACKING_SHOW, uint32(IsWarTime() && GetAttackerTeam() == TEAM_ALLIANCE ? 1 : 0));
+    sWorld->setWorldState(TB_WS_HORDE_ATTACKING_SHOW, uint32(IsWarTime() && GetAttackerTeam() == TEAM_HORDE ? 1 : 0));
+    sWorld->setWorldState(TB_WS_TIME_NEXT_BATTLE, uint32(!IsWarTime() ? m_Timer : 0));
+    sWorld->setWorldState(TB_WS_TIME_NEXT_BATTLE_SHOW, uint32(!IsWarTime() ? 1 : 0));
 
     // Tol Barad
     for (uint8 team = 0; team < 2; team++)
@@ -473,7 +473,7 @@ void BattlefieldTB::UpdateNPCsAndGameObjects()
         if (GameObject* gameobject = GetGameObject(guid))
             gameobject->Delete();
     TemporaryGOs.clear();
-    
+
     // Tol Barad gates - closed during warmup
     if (GameObject* gates = GetGameObject(TBGatesGUID))
         gates->SetGoState(GetState() == BATTLEFIELD_WARMUP ? GO_STATE_READY : GO_STATE_ACTIVE);
@@ -533,7 +533,7 @@ void BattlefieldTB::UpdateNPCsAndGameObjects()
             if (Creature* creature = SpawnCreature(entry, TBQuestInfantrySpawnData[i], GetDefenderTeam()))
                 TemporaryNPCs.insert(creature->GetGUID());
         }
-            
+
         for (uint8 i = 0; i < TB_GUARDS_MAX; i++)
             if (Creature* creature = SpawnCreature(GetDefenderTeam() == TEAM_ALLIANCE ? NPC_BARADIN_GUARD : NPC_HELLSCREAMS_SENTRY, GuardNPCSpawns[i], GetDefenderTeam()))
                 TemporaryNPCs.insert(creature->GetGUID());
